server/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["AG <goebel.aaron@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.67.0"
uvicorn = "^0.14.0"
gunicorn = "^20.1.0"
alembic = "^1.6.5"
psycopg2 = "^2.9.1"
black = "^21.7b0"
pandas = "^1.3.1"
transitions = "^0.8.8"
starlette-context = "^0.3.3"
web3 = "^5.22.0"
PyYAML = "^5.4.1"
ipfshttpclient = "^0.7.0"
google-cloud-storage = "^1.41.1"
smart-open = {version = "^5.1.0", extras = ["gcs"]}
python-multipart = "^0.0.5"
gcloud-aio-storage = "^6.1.0"
certifi = "^2021.5.30"
gcloud-aio-auth = "^3.6.1"
gino-starlette = "^0.1.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
requests = "^2.26.0"
ipdb = "^0.13.9"
jupyter = "^1.0.0"
nest-asyncio = "^1.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins."api.modules"]
users= "api.views.users"
disputes = "api.views.disputes"
indexers = "api.views.indexers"
divergent_blocks = "api.views.divergent_blocks"
upload = "api.transport.upload"
list = "api.transport.list"
